+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Generator for Kohana - Changelog
 ================================
 
+Version 1.0
+-----------
+
+- Various tweaks and code cleanup for Kohana 3.3.0 release
 - Added option for generators to define custom base paths for destination files
   outside of the default APPPATH or MODPATH.